fix secureStorage initialisation issue
